third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/paint-order-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'paint-order' to the 'fill' keyword: fill
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'paint-order' to the 'stroke' keyword: stroke
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'paint-order' to the 'markers' keyword: markers
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
Harness: the test ran to completion.
